src/main/java/ru/netology/manager/Manager.java
package ru.netology.manager;

import lombok.Data;
import ru.netology.domain.Book;
import ru.netology.domain.Product;
import ru.netology.domain.Smartphone;
import ru.netology.repository.Repository;


@Data
public class Manager {
    Repository repository_M = new Repository();
    Repository repository=new Repository();

    public void save(Product product) {
        repository_M.save(product);
    }

    public Product[] searchBy(String text) {
        Product[] result = new Product[0];
        for (Product product : repository_M.findAll())
            if (matches(product, text)) {
                Product[] tmp = new Product[result.length + 1];
                System.arraycopy(result, 0, tmp, 0, tmp.length - 1);
                tmp[tmp.length - 1] = product;
                result = tmp;
            }
        return result;
    }

    public boolean matches(Product product, String search) {
        if (product instanceof Book) {
            Book book = (Book) product;
            if (book.getName().equalsIgnoreCase(search)) {
                return true;
            }
            if (book.getAuthor().equalsIgnoreCase(search)) {
                return true;
            }
            // return false;
        }
        if (product instanceof Smartphone) {
            Smartphone smartphone = (Smartphone) product;
            if (smartphone.getName().equalsIgnoreCase(search)) {
                return true;
            }
            return smartphone.getManufacturer().equalsIgnoreCase(search);
        }
        return false;
    }

}